+++ include/UIElement.h
@@ -8,9 +8,9 @@ namespace UI {
 class UIElement : public GameObject {
 
 public:
-    void Scale(int8_t scalar);
-    void Translate(int8_t delta_x, int8_t delta_y);
-    void PlaceAt(uint16_t _x, uint16_t _y);
+    void Scale(int32_t scalar);
+    void Translate(int32_t delta_x, int32_t delta_y);
+    void PlaceAt(uint32_t _x, uint32_t _y);
     
     SDL_Rect* GetRect();
     SDL_Texture** GetTexture();
